Refactor `Config` type out of `config-utils`
